# packages/contracts/openapi/v1.yaml
openapi: 3.0.3
info:
  title: 360Solve Connector API
  description: Multi-tenant ERP Integration Platform API
  version: 1.0.0
  contact:
    name: 360Solve Support
    email: support@360solve.com
servers:
  - url: http://localhost:3001/api/v1
    description: Local development server
security:
  - bearerAuth: []
  - tenantId: []
paths:
...
  /workflows:
    get:
      tags:
        - Workflows
      summary: List workflows
      description: Get paginated list of workflows for tenant
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, archived]
      responses:
        200:
          description: List of workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'

    post:
      tags:
        - Workflows
      summary: Create workflow
      description: Create a new workflow definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        201:
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
  /workflows/{id}:
    get:
      tags:
        - Workflows
      summary: Get workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        404:
          description: Workflow not found

    put:
      tags:
        - Workflows
      summary: Update workflow
      description: Update workflow (creates new version)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
      responses:
        200:
          description: Workflow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'

    delete:
      tags:
        - Workflows
      summary: Delete workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Workflow deleted
...
  /connectors:
    get:
      tags:
        - Connectors
      summary: List Mini Connectors
      responses:
        200:
          description: List of connectors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorListResponse'

    post:
      tags:
        - Connectors
      summary: Create new connector
      description: Generate API key for new Mini Connector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectorRequest'
      responses:
        201:
          description: Connector created with API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorWithKeyResponse'
  /connectors/{id}:
    delete:
      tags:
        - Connectors
      summary: Revoke connector
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Connector revoked

  # AI endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
  schemas:
...
    CreateWorkflowRequest:
      type: object
      required:
        - name
        - definition
      properties:
        name:
          type: string
        description:
          type: string
        definition:
          $ref: '#/components/schemas/WorkflowDefinition'
    UpdateWorkflowRequest:
      type: object
      required:
        - name
        - definition
      properties:
        name:
          type: string
        description:
          type: string
        definition:
          $ref: '#/components/schemas/WorkflowDefinition'
    WorkflowDefinition:
      type: object
      required:
        - nodes
        - connections
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowNode'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowConnection'
        schedule:
          $ref: '#/components/schemas/ScheduleConfig'
    WorkflowNode:
      type: object
      required:
        - id
        - type
        - label
      properties:
        id:
          type: string
        type:
          type: string
          enum: [source, transform, destination]
        label:
          type: string
        config:
          type: object
    WorkflowConnection:
      type: object
      required:
        - from
        - to
      properties:
        from:
          type: string
        to:
          type: string
    ScheduleConfig:
      type: object
      properties:
        type:
          type: string
          enum: [manual, interval, cron]
        expression:
          type: string
        interval:
          type: integer
          description: Seconds between runs
    WorkflowResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            description:
              type: string
            version:
              type: integer
            hash:
              type: string
            status:
              type: string
              enum: [draft, active, archived]
            definition:
              $ref: '#/components/schemas/WorkflowDefinition'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    WorkflowListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowListItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    WorkflowListItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
        lastRun:
          type: string
          format: date-time
        nextRun:
          type: string
          format: date-time
        successRate:
          type: number
        totalRuns:
          type: integer
        createdAt:
          type: string
          format: date-time

    # Execution schemas
...
    CreateConnectorRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        os:
          type: string
          enum: [windows, linux, macos]
    ConnectorWithKeyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            apiKey:
              type: string
            downloadUrl:
              type: string
            setupInstructions:
              type: string
    ConnectorListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/ConnectorListItem'
    ConnectorListItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [online, offline, unknown]
        ipAddress:
          type: string
        version:
          type: string
        lastSeen:
          type: string
          format: date-time
        activeWorkflows:
          type: integer

    # Common schemas
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...